# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
  secret-service_internal_domain.AccessGrant:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      granted_by:
        type: string
      id:
        type: string
      secret_id:
        type: string
      user_id:
        type: string
    type: object
...
  secret-service_internal_domain.Secret:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      environment:
        $ref: '#/definitions/secret-service_internal_domain.SecretEnvironment'
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        $ref: '#/definitions/secret-service_internal_domain.SecretStatus'
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  secret-service_internal_domain.SecretEnvironment:
    enum:
    - development
    - staging
    - production
    type: string
    x-enum-varnames:
    - EnvDevelopment
    - EnvStaging
    - EnvProduction
  secret-service_internal_domain.SecretStatus:
    enum:
    - active
    - revoked
    type: string
    x-enum-varnames:
    - SecretStatusActive
    - SecretStatusRevoked
  secret-service_internal_domain.SecretVersion:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_current:
        type: boolean
      secret_id:
        type: string
      version:
        type: integer
    type: object
...
  secret-service_internal_dto.CreateSecretRequest:
    properties:
      description:
        type: string
      environment:
        type: string
      expires_at:
        type: string
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      value:
        type: string
    type: object
...
  secret-service_internal_dto.GrantAccessRequest:
    properties:
      expires_at:
        type: string
      user_id:
        type: string
    type: object
...
  secret-service_internal_dto.RollbackSecretRequest:
    properties:
      version:
        type: integer
    type: object
  secret-service_internal_dto.RotateSecretRequest:
    properties:
      value:
        type: string
    type: object
  secret-service_internal_dto.SecretValueResponse:
    properties:
      secret_id:
        type: string
      value:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Centralized API key storage and management service for development
    teams.
  title: Secret Service API
  version: "1.0"
paths:
...
  /projects/{projectID}/secrets:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.Secret'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List secrets by project
      tags:
      - secrets
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Secret name, description and value
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret-service_internal_domain.Secret'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create secret
      tags:
      - secrets
  /projects/{projectID}/secrets/{secretID}/grants:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.AccessGrant'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List access grants for a secret
      tags:
      - secrets
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      - description: User ID and optional expiration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.GrantAccessRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Grant access to secret
      tags:
      - secrets
  /projects/{projectID}/secrets/{secretID}/grants/{userID}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Revoke user access to secret
      tags:
      - secrets
  /projects/{projectID}/secrets/expiring:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Number of days to look ahead (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.Secret'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List secrets expiring soon
      tags:
      - secrets
...
  /secrets/{secretID}/revoke:
    post:
      parameters:
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Revoke secret
      tags:
      - secrets
  /secrets/{secretID}/rollback:
    post:
      consumes:
      - application/json
      parameters:
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      - description: Target version number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.RollbackSecretRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Rollback secret to a previous version
      tags:
      - secrets
  /secrets/{secretID}/rotate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      - description: New secret value
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.RotateSecretRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Rotate secret value
      tags:
      - secrets
  /secrets/{secretID}/value:
    get:
      parameters:
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret-service_internal_dto.SecretValueResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get secret value
      tags:
      - secrets
  /secrets/{secretID}/versions:
    get:
      parameters:
      - description: Secret ID
        in: path
        name: secretID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.SecretVersion'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List secret version history
      tags:
      - secrets
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"